+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Add support to parse formulas with all whitespace characters (controlled by config param `allowAllWhitespace`). (#898)
+- Add support to parse formulas with all whitespace characters (controlled by config param `ignoreWhiteSpace`). (#898)
 
 ### Changed
 - Removed `gpu.js` dependency and its use. (#812)